Separate Frame from Resource folder to support native MacOS port.
Fixes https://github.com/audetto/AppleWin/issues/55
