+++ source/frontends/common2/commonframe.cpp
@@ -1,101 +1,23 @@
 #include "StdAfx.h"
 #include "frontends/common2/commonframe.h"
 #include "frontends/common2/utils.h"
-#include "frontends/common2/fileregistry.h"
 #include "linux/resources.h"
-#include "linux/context.h"
 
 #include <sys/stat.h>
 #include <fcntl.h>
 #include <unistd.h>
-#include <libgen.h>
-
-#ifdef __APPLE__
-#include "mach-o/dyld.h"
-#endif
 
 #include "Log.h"
-#include "Core.h"
-#include "config.h"
-
-namespace
-{
-
-  bool dirExists(const std::string & folder)
-  {
-    struct stat stdbuf;
-
-    if (stat(folder.c_str(), &stdbuf) == 0 && S_ISDIR(stdbuf.st_mode))
-    {
-      return true;
-    }
-    else
-    {
-      return false;
-    }
-  }
-
-  std::string getResourcePath(const std::string & target)
-  {
-    std::vector<std::string> paths;
-
-    char self[1024] = {0};
-
-#ifdef __APPLE__
-    uint32_t size = sizeof(self);
-    const int ch = _NSGetExecutablePath(self, &size);
-#else
-    const int ch = readlink("/proc/self/exe", self,  sizeof(self));
-#endif
-
-    if (ch != -1)
-    {
-      const char * path = dirname(self);
-
-      // case 1: run from the build folder
-      paths.emplace_back(std::string(path) + '/'+ ROOT_PATH);
-      // case 2: run from the installation folder
-      paths.emplace_back(std::string(path) + '/'+ SHARE_PATH);
-    }
-
-    // case 3: use the source folder
-    paths.emplace_back(CMAKE_SOURCE_DIR);
-
-    for (const std::string & path : paths)
-    {
-      char * real = realpath(path.c_str(), nullptr);
-      if (real)
-      {
-        const std::string resourcePath = std::string(real) + target;
-        free(real);
-        if (dirExists(resourcePath))
-        {
-          return resourcePath;
-        }
-      }
-    }
-
-    throw std::runtime_error("Cannot found the resource path: " + target);
-  }
-
-}
 
 namespace common2
 {
-
-  CommonFrame::CommonFrame()
-    : myResourcePath(getResourcePath("/resource/"))
-  {
-    // should this go down to LinuxFrame (maybe Initialisation?)
-    g_sProgramDir = getResourcePath("/bin/");
-  }
 
   BYTE* CommonFrame::GetResource(WORD id, LPCSTR lpType, DWORD expectedSize)
   {
     myResource.clear();
 
     const std::string & filename = getResourceName(id);
-    const std::string path = myResourcePath + filename;
+    const std::string path = getResourcePath(filename);
 
     const int fd = open(path.c_str(), O_RDONLY);
 
@@ -131,11 +53,6 @@ namespace common2
     if (resource == "CHARSET8C") return "CHARSET8C.bmp";
 
     return resource;
-  }
-
-  std::string CommonFrame::Video_GetScreenShotFolder()
-  {
-    return GetHomeDir() + "/Pictures/";
   }
 
 }
