+++ source/frontends/common2/commonframe.h
@@ -10,16 +10,13 @@ namespace common2
   class CommonFrame : public LinuxFrame
   {
   public:
-    CommonFrame();
-
     BYTE* GetResource(WORD id, LPCSTR lpType, DWORD expectedSize) override;
 
-    std::string Video_GetScreenShotFolder() override;
-
   protected:
+    virtual std::string getResourcePath(const std::string & filename) = 0;
+
     static std::string getBitmapFilename(const std::string & resource);
 
-    const std::string myResourcePath;
     std::vector<BYTE> myResource;
   };
