+++ source/frontends/libretro/retroframe.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "frontends/common2/commonframe.h"
+#include "frontends/common2/gnuframe.h"
 
 #include <memory>
 #include <vector>
@@ -8,7 +9,7 @@
 namespace ra2
 {
 
-  class RetroFrame : public common2::CommonFrame
+  class RetroFrame : public virtual common2::CommonFrame, public common2::GNUFrame
   {
   public:
     RetroFrame();
