+++ source/frontends/sdl/renderer/sdlrendererframe.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "frontends/sdl/sdlframe.h"
+#include "frontends/common2/gnuframe.h"
 
 namespace common2
 {
@@ -10,7 +11,7 @@ namespace common2
 namespace sa2
 {
 
-  class SDLRendererFrame : public SDLFrame
+  class SDLRendererFrame : public SDLFrame, public common2::GNUFrame
   {
   public:
     SDLRendererFrame(const common2::EmulatorOptions & options);
