+++ source/frontends/sdl/sdlframe.cpp
@@ -178,7 +178,7 @@ namespace sa2
 
   void SDLFrame::SetApplicationIcon()
   {
-    const std::string path = myResourcePath + "APPLEWIN.ICO";
+    const std::string path = getResourcePath("APPLEWIN.ICO");
     std::shared_ptr<SDL_Surface> icon(IMG_Load(path.c_str()), SDL_FreeSurface);
     if (icon)
     {
@@ -194,7 +194,7 @@ namespace sa2
   void SDLFrame::GetBitmap(LPCSTR lpBitmapName, LONG cb, LPVOID lpvBits)
   {
     const std::string filename = getBitmapFilename(lpBitmapName);
-    const std::string path = myResourcePath + filename;
+    const std::string path = getResourcePath(filename);
 
     std::shared_ptr<SDL_Surface> surface(SDL_LoadBMP(path.c_str()), SDL_FreeSurface);
     if (surface)
